# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_handler.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/internal_handler.APIError'
      meta:
        $ref: '#/definitions/internal_handler.Meta'
      success:
        type: boolean
    type: object
  internal_handler.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_handler.deviceTokenResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      platform:
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
...
  internal_handler.favoriteResponse:
    properties:
      bathhouse_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
...
  internal_handler.registerDeviceTokenRequest:
    properties:
      platform:
        type: string
      token:
        type: string
    type: object
...
  internal_handler.statusResponse:
    properties:
      status:
        type: string
    type: object
...
  internal_handler.toggleFavoriteResponse:
    properties:
      is_favorite:
        type: boolean
    type: object
...
info:
  contact: {}
  description: API for the Bani bathhouse marketplace platform.
  title: Bani API
  version: "1.0"
paths:
...
  /bathhouses/{id}/favorite:
    post:
      description: Adds or removes a bathhouse from the user's favorites
      parameters:
      - description: Bathhouse ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.toggleFavoriteResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Toggle favorite
      tags:
      - favorites
...
  /device-tokens:
    post:
      consumes:
      - application/json
      description: Registers a device token for push notifications (FCM, APNs)
      parameters:
      - description: Device token and platform
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.registerDeviceTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.deviceTokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Register device token
      tags:
      - device-tokens
  /device-tokens/{id}:
    delete:
      description: Removes a registered device token, disabling push notifications
        for that device
      parameters:
      - description: Device token ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.statusResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Delete device token
      tags:
      - device-tokens
...
  /my/favorites:
    get:
      description: Returns paginated list of the user's favorite bathhouses
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handler.favoriteResponse'
                  type: array
                meta:
                  $ref: '#/definitions/internal_handler.Meta'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: List favorites
      tags:
      - favorites
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"